Codegen: Rename `codegen.py` -> `codegenerator.py`
This avoids namespace clashes with the `codegen` package.
